# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText - OpenAPI 3.0
  description: |-
    This is the OpenAPI specification for the WASAText web-app.
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /conversations/{conversationId}/messages/{messageId}/reactions:
    parameters:
      - name: conversationId
        in: path
        description: |-
            The id of the conversation
        required: true
        schema:
          type: integer
          minimum: 1
          example: 42
      - name: messageId
        in: path
        description: The id of the message to react to
        required: true
        schema:
          type: integer
          minimum: 1
          example: 18284
    put:
      tags: ["reactions"]
      operationId: commentMessage
      summary: Add or update reaction to a message (comment)
      description: |-
        Adds or updates a reaction (emoticon) to the specified message.
        Each user can have only ONE reaction per message.
        If the user already has a reaction on this message, it will be replaced,
        otherwise a new reaction is created.
        The user is identified by the Bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionEmoticon"
      responses:
        "200":
          description: |-
              Reaction added or updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionCreated"
        "400":
          description: |-
              Bad request - Invalid emoticon
        "401":
          description: |-
              Unauthorized
        "403":
          description: |-
              Forbidden - Not a participant in this conversation
        "404":
          description: Message not found
    delete:
      tags: ["reactions"]
      operationId: uncommentMessage
      summary: Remove reaction from message (uncomment)
      description: |-
        Removes the user's reaction (emoticon) from the specified message.
        The user is identified by the Bearer token.
      responses:
        "204":
          description: Reaction removed successfully
        "400":
          description: Bad request - Invalid message or conversation ID
        "401":
          description: Unauthorized
        "404":
          description: Message or reaction not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        User identifier returned from login.
        Use the identifier as the bearer token.
  schemas:
...
    ReactionEmoticon:
      description: |-
          This schema represents the emoticon of a reaction.
      type: object
      required:
        - emoticon
      properties:
        emoticon:
          type: string
          pattern:  '^.*$'
          example:  "😄"
          minLength: 1
          maxLength: 10
          description: |-
              Emoticon to use as reaction (emoji or text)
    ReactionCreated:
      description: |-
          A reaction just created.
      type: object
      properties:
        username:
          description: |-
              The username of the user who sent a reaction.
          type: string
          example: "Maria"
        emoticon:
          description: |-
              The emoticon of the reaction.
          type: string
          example: "😄"
        messageId:
          description: |-
              The id of the message of the reaction.
          type: integer
          example: 18284
...
security:
  - bearerAuth: []